requirements.txt
adal==0.4.5
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.22.0
azure==2.0.0
azure-batch==3.0.0
azure-common==1.1.6
azure-datalake-store==0.0.12
azure-graphrbac==0.30.0
azure-keyvault==0.3.5
azure-mgmt==1.0.0
azure-mgmt-authorization==0.30.0
azure-mgmt-batch==4.0.0
azure-mgmt-cdn==0.30.3
azure-mgmt-cognitiveservices==1.0.0
azure-mgmt-compute==1.0.0
azure-mgmt-containerregistry==0.2.1
azure-mgmt-datalake-analytics==0.1.6
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.1.6
azure-mgmt-devtestlabs==2.0.0
azure-mgmt-dns==1.0.1
azure-mgmt-documentdb==0.1.3
azure-mgmt-iothub==0.2.2
azure-mgmt-keyvault==0.31.0
azure-mgmt-logic==2.1.0
azure-mgmt-monitor==0.2.1
azure-mgmt-network==1.0.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-rdbms==0.1.0
azure-mgmt-redis==4.1.0
azure-mgmt-resource==1.1.0
azure-mgmt-scheduler==1.1.2
azure-mgmt-sql==0.5.3
azure-mgmt-storage==1.0.0
azure-mgmt-trafficmanager==0.30.0
azure-mgmt-web==0.32.0
azure-nspkg==2.0.0
azure-servicebus==0.21.1
azure-servicefabric==5.6.130
azure-servicemanagement-legacy==0.20.6
azure-storage==0.34.3
billiard==3.3.0.23
celery==3.1.20
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==1.9
Django==1.11.2
djangorestframework==3.6.3
idna==2.5
isodate==0.5.4
keyring==10.4.0
kombu==3.0.37
msrest==0.4.11
msrestazure==0.4.11
oauthlib==2.0.2
pycparser==2.17
PyJWT==1.5.2
pymongo==3.4.0
python-dateutil==2.6.0
pytz==2017.2
pywin32-ctypes==0.1.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
SQLAlchemy==1.1.11
urllib3==1.21.1
vine==1.1.3
